requirements.txt
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
autopep8==2.0.4
Babel==2.14.0
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.3.0
debugpy==1.8.1
decorator==5.1.1
Django==5.0.4
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.3.1
django-countries==7.6.1
django-enumfield==3.1
django-enumfields==2.1.1
django-extensions==3.2.3
django-idempotency-key==1.3.0
django-iranian-cities==1.0.2
django-modeltranslation==0.18.11
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
drf-yasg==1.21.7
ecdsa==0.18.0
enum34==1.1.10
exceptiongroup==1.2.0
executing==2.0.1
flake8==7.0.0
flatbuffers==24.3.7
geonamescache==2.0.0
idna==3.6
imagesize==1.4.1
inflection==0.5.1
ipykernel==6.29.4
ipython==8.23.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kombu==5.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.28
pika==1.3.2
platformdirs==4.2.0
progressbar2==4.4.2
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pycodestyle==2.11.1
pydantic==2.6.3
pydantic_core==2.16.3
pydotplus==2.0.2
PyEventDrivenTools==0.1.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyocclient==0.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-utils==3.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
django-cities-light==3.10.1
